+++ src/calc_path/calc_path.cuh
@@ -10,11 +10,11 @@
 #include <iostream>
 #include <bitset>
 
-__global__ void matMult ( int * pix_cost, long*  agg_cost,  size_t rows, size_t cols );
-__global__ void optimised_concatResCUDA ( long*  agg_cost, int* res,  size_t rows, size_t cols );
+__global__ void matMult ( int * pix_cost, int*  agg_cost,  size_t rows, size_t cols );
+__global__ void optimised_concatResCUDA ( int*  agg_cost, int* res,  size_t rows, size_t cols );
 
-__global__ void optimized_matMult_LEFT ( int * pix_cost, long* agg_cost,  size_t rows, size_t cols );
-__global__ void optimized_matMult_RIGHT ( int * pix_cost, long* agg_cost,  size_t rows, size_t cols );
-__global__ void optimized_matMult_TOP ( int * pix_cost, long* agg_cost,  size_t rows, size_t cols );
+__global__ void optimized_matMult_LEFT ( int * pix_cost, int* agg_cost,  size_t rows, size_t cols );
+__global__ void optimized_matMult_RIGHT ( int * pix_cost, int* agg_cost,  size_t rows, size_t cols );
+__global__ void optimized_matMult_TOP ( int * pix_cost, int* agg_cost,  size_t rows, size_t cols );
 __global__ void clearResCUDA ( int * res,  size_t rows, size_t cols );
